FINALLY FUNCTIONAL
FINALLY FIXED IT!!!! All it required was moving the JS script in the head, down to the body of the HTML - researched, and determined that JS was trying to load before the HTML has loaded, therefore could not read any element. Moved it to the bottom of the HTML, problem solved. Very happy.